Django/myproject/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-seed==0.3.1
djangorestframework==3.14.0
Faker==22.6.0
python-dateutil==2.8.2
pytz==2023.4
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2023.4
